+++ src/test/java/com/microsoft/jenkins/appservice/commands/GitDeployCommandTest.java
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See License.txt in the project root for
  * license information.
  */
-package com.microsoft.jenkins.appservice.test;
+package com.microsoft.jenkins.appservice.commands;
 
 import hudson.FilePath;
 import hudson.remoting.VirtualChannel;
