+++ src/test/resources/com/microsoft/jenkins/appservice/integration/sample-java-app/app.war
