+++ src/test/resources/com/microsoft/jenkins/appservice/integration/sample-nodejs-app/index.js
