+++ src/test/resources/com/microsoft/jenkins/appservice/integration/sample-nodejs-app/package.json
