+++ src/test/resources/com/microsoft/jenkins/appservice/integration/sample-python-app/main.py
