+++ src/test/resources/com/microsoft/jenkins/appservice/integration/sample-python-app/web.3.4.config
